requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0.post1
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.6.0
psutil==5.9.8
starlette==0.35.1
google-cloud-translate==3.15.0
black==22.3.0
isort==5.10.1
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0